Добавлена поддержка float значений [0.0.3]

+++ Memory/GlobalRegister.cs
@@ -2,7 +2,7 @@
 {
     internal class GlobalRegister
     {
-        public static int Accumulator = 0;
+        public static float Accumulator = 0;
         public static int InstructionPointer = 0;
         public static string RunningFunction = "";
     }
